requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.4.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
Werkzeug==2.1.2
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1
gunicorn==20.1.0
